requirements.txt
streamlit
python-docx
pandas
deepl
python-dotenv
streamlit-option-menu